test(feishu): split media resource tests
